src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.6
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
flake8==7.1.1
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==6.0.0
loguru==0.7.3
mccabe==0.7.0
multidict==6.3.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pyTelegramBotAPI==4.25.0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
win32_setctime==1.2.0
yarl==1.18.3
